pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zte-f680-mcp"
version = "0.3.0"
description = "MCP server to manage NAT / port forwarding on ZTE ZXHN F680 routers"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
  { name = "Alberto Diaz" },
]
keywords = [
  "mcp",
  "model-context-protocol",
  "zte",
  "f680",
  "zxhn",
  "router",
  "nat",
  "port-forwarding",
  "claude",
  "llm",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: System Administrators",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: System :: Networking",
  "Topic :: Internet",
]
dependencies = [
  "mcp[cli]>=1.0.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
]

[project.scripts]
zte-f680-mcp = "zte_f680_mcp.server:main"

[project.urls]
Homepage = "https://github.com/Picaresco/MCP-ZTE-F680"
Repository = "https://github.com/Picaresco/MCP-ZTE-F680"
Issues = "https://github.com/Picaresco/MCP-ZTE-F680/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/zte_f680_mcp"]

[tool.hatch.build.targets.sdist]
include = [
  "src/zte_f680_mcp",
  "README.md",
  "LICENSE",
  "pyproject.toml",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "router: tests que requieren router ZTE real (desactivados por defecto)",
]
